fix: removed unnecessary codes
