+++ frontend/src/assets/css/ProductCard.css
@@ -16,6 +16,11 @@
   flex-basis: 30%;
 }
 
+.overlay {
+  text-align: center;
+  font-weight: 600;
+}
+
 .product-img--container .image {
   height: 8rem;
   object-fit: contain;
